Move options parsing stuff to apiman-common-config
