+++ gateway/platforms/servlet/pom.xml
@@ -28,6 +28,10 @@
       <groupId>${project.groupId}</groupId>
       <artifactId>apiman-gateway-engine-policies</artifactId>
     </dependency>
+    <dependency>
+      <groupId>${project.groupId}</groupId>
+      <artifactId>apiman-common-config</artifactId>
+    </dependency>
     
     <!-- Third Party Dependencies -->
     <dependency>
